+++ CHANGES.rst
@@ -5,3 +5,6 @@ Changelog (``cosmology.compat``)
 ------------------
 
 - Initial commit (#1) [nstarman]
+
+- Add tests and ensure that the astropy cosmology class is a conformatn wrapper.
+  (#2) [nstarman]